Finish affectation

+++ src/main/java/com/example/applicationmaven1/repositories/ReservationRepository.java
@@ -3,5 +3,5 @@ package com.example.applicationmaven1.repositories;
 import com.example.applicationmaven1.Entities.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 }
